+++ apps/improve/src/app/app.module.ts
@@ -1,11 +1,10 @@
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
+import { SharedModule } from './modules/shared/public-api';
 import { AppRoutingModule } from './app-routing.module';
-
 import { AppComponent } from './components/app/app.component';
 import { ClassSelectorComponent } from './components/class-selector/class-selector.component';
-import { SharedModule } from './modules/shared/shared.module';
 
 @NgModule({
   declarations: [AppComponent, ClassSelectorComponent],