+++ apps/improve/src/app/modules/classes/classes.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
 
+import { SharedModule } from '../shared/public-api';
 import { ClassesRoutingModule } from './classes-routing.module';
 import { AddComponent } from './pages/add/add.component';
-import { ReactiveFormsModule } from '@angular/forms';
-import { SharedModule } from '../shared/shared.module';
 import { ListComponent } from './pages/list/list.component';
 
 @NgModule({