+++ apps/improve/src/app/modules/database/public-api.ts
@@ -3,5 +3,7 @@ export { DatabaseModule } from './database.module';
 export { DatabaseService } from './services/database/database.service';
 export { ThemesService } from './services/themes/themes.service';
 export { UsedThemesService } from './services/used-themes/used-themes.service';
+export { ArticulationsService } from './services/articulations/articulations.service';
+export { UsedArticulationsService } from './services/used-articulations/used-articulations.service';
 export { ClassesService } from './services/classes/classes.service';
 export { CurrentClassService } from './services/current-class/current-class.service';