+++ apps/improve/src/app/modules/shared/public-api.ts
@@ -1,0 +1,13 @@
+export { SharedModule } from './shared.module';
+
+export { FeatherIconComponent } from './components/feather-icon/feather-icon.component';
+export { LayoutComponent } from './components/layout/layout.component';
+export { LayoutWithButtonsComponent } from './components/layout-with-buttons/layout-with-buttons.component';
+
+export {
+  IItemsService,
+  ITEMS_SERVICE,
+  USED_SERVICE,
+  IUsedService,
+  ItemsHelperService,
+} from './services/items-helper/items-helper.service';