+++ apps/improve/src/app/modules/shared/services/items-helper/items-helper.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ThemesHelperService } from './themes-helper.service';
+import { ItemsHelperService } from './items-helper.service';
 
-describe('ThemesHelperService', () => {
-  let service: ThemesHelperService;
+describe('ItemsHelperService', () => {
+  let service: ItemsHelperService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(ThemesHelperService);
+    service = TestBed.inject(ItemsHelperService);
   });
 
   it('should be created', () => {